+++ worker/src/user_api/user.ts
@@ -10,8 +10,7 @@ import { sendMail } from "../mails_api/send_mail_api";
 export default {
     verifyCode: async (c: Context<HonoCustomType>) => {
         const { email, cf_token } = await c.req.json();
-        const lang = c.get("lang") || c.env.DEFAULT_LANG;
-        const msgs = i18n.getMessages(lang);
+        const msgs = i18n.getMessagesbyContext(c);
         // check cf turnstile
         try {
             await checkCfTurnstile(c, cf_token);
@@ -27,6 +26,17 @@ export default {
             && !settings.mailAllowList.includes(mailDomain)
         ) {
             return c.text(`${msgs.UserMailDomainMustInMsg} ${JSON.stringify(settings.mailAllowList, null, 2)}`, 400)
+        }
+        // check email regex
+        if (settings.enableEmailCheckRegex && settings.emailCheckRegex) {
+            try {
+                const regex = new RegExp(settings.emailCheckRegex);
+                if (!regex.test(email)) {
+                    return c.text(`${msgs.UserEmailNotMatchRegexMsg}: /${settings.emailCheckRegex}/`, 400)
+                }
+            } catch (e) {
+                console.error("Failed to check user email regex", e);
+            }
         }
         if (!settings.verifyMailSender) {
             return c.text(msgs.VerifyMailSenderNotSetMsg, 400)
@@ -61,8 +71,7 @@ export default {
     register: async (c: Context<HonoCustomType>) => {
         const value = await getJsonSetting(c, CONSTANTS.USER_SETTINGS_KEY);
         const settings = new UserSettings(value)
-        const lang = c.get("lang") || c.env.DEFAULT_LANG;
-        const msgs = i18n.getMessages(lang);
+        const msgs = i18n.getMessagesbyContext(c);
         // check enable
         if (!settings.enable) {
             return c.text(msgs.UserRegistrationDisabledMsg, 403);
@@ -83,6 +92,17 @@ export default {
             && !settings.mailAllowList.includes(mailDomain)
         ) {
             return c.text(`${msgs.UserMailDomainMustInMsg} ${JSON.stringify(settings.mailAllowList, null, 2)}`, 400)
+        }
+        // check email regex
+        if (settings.enableEmailCheckRegex && settings.emailCheckRegex) {
+            try {
+                const regex = new RegExp(settings.emailCheckRegex);
+                if (!regex.test(email)) {
+                    return c.text(`${msgs.UserEmailNotMatchRegexMsg}: /${settings.emailCheckRegex}/`, 400)
+                }
+            } catch (e) {
+                console.error("Failed to check user email regex", e);
+            }
         }
         // check code
         if (settings.enableMailVerify) {
@@ -154,8 +174,7 @@ export default {
     },
     login: async (c: Context<HonoCustomType>) => {
         const { email, password } = await c.req.json();
-        const lang = c.get("lang") || c.env.DEFAULT_LANG;
-        const msgs = i18n.getMessages(lang);
+        const msgs = i18n.getMessagesbyContext(c);
         if (!email || !password) return c.text(msgs.InvalidEmailOrPasswordMsg, 400);
         const { id: user_id, password: dbPassword } = await c.env.DB.prepare(
             `SELECT id, password FROM users where user_email = ?`